Bugfix: Not loading new entries
The abckend takes a while until new entries are available. However, cgmAccess logic  only queried each time frame once. If nothing was returned, we would not query a second time.
This bugfix only updates "latest_query_time" if new values were found and sets it to the latest timestamp of the retuned values